chore: move install to Registry
This is an effort to consolidate all handling of browser binaries in a single place.
