+++ src/cli/driver.ts
@@ -22,12 +22,10 @@ import { BrowserType } from '../client/browserType';
 import { LaunchServerOptions } from '../client/types';
 import { DispatcherConnection } from '../dispatchers/dispatcher';
 import { PlaywrightDispatcher } from '../dispatchers/playwrightDispatcher';
-import { installBrowsersWithProgressBar } from '../install/installer';
 import { Transport } from '../protocol/transport';
 import { PlaywrightServer } from '../remote/playwrightServer';
 import { createPlaywright } from '../server/playwright';
 import { gracefullyCloseAll } from '../server/processLauncher';
-import { BrowserName } from '../utils/registry';
 
 export function printApiJson() {
   // Note: this file is generated by build-playwright-driver.sh
@@ -65,8 +63,4 @@ export async function launchBrowserServer(browserName: string, configFile?: stri
   const browserType = (playwright as any)[browserName] as BrowserType;
   const server = await browserType.launchServer(options);
   console.log(server.wsEndpoint());
-}
-
-export async function installBrowsers(browserNames?: BrowserName[]) {
-  await installBrowsersWithProgressBar(browserNames);
 }
